Menu handling cleanup.
